feat(viewport): give 1 extra column for empty eol (end-of-line) (#424)
The eol (end-of-line) uses 0-width cells in terminal in most cases, and cursor will never need to move to it (in normal mode). While in insert mode cursor need to move to it and thus append more chars at the line end.
This PR gives 1 extra column for empty eol when searching viewport anchor.