test(evloop): add "run_with_mock_operations" for operation based loop run (#612)
This PR provides a new way to run event loop with mocked operations, instead of crossterm's events stream. Since keyboard event based stream is not efficient, for example if we want to test user inserts a text payload with N characters, we have to mock N more events (each event for exactly one char). But with operation based mocking, we can directly use a higher level operation. For this example, we can just use one "Operation::CursorInsert" to complete the task.