perf(js): avoid duplicate loop tick when waiting for dynamic import loaders (#654)
This PR is a follow up of #631 , it avoids some duplicated event loop ticks in dynamic import stage. In previous implementation, it always requires another tick if "module_map.pending" is not empty. In this PR, it only requires tick when "module_map.pending.len() > pending_imports.len()". Because the "async loader" is handled by tokio runtime, and it actually doesn't need an event loop tick to run.

BTW, #651 is also a follow up of #631 , it rewrites the "JsFuture" and messages between tokio runtime and js runtime event loop, to make the code base cleaner.